requirements.txt
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
# NOTE: TensorFlow does not currently provide wheels in the 2.12–2.15 range
# for Python 3.12 on Windows. We therefore depend on a more recent version
# for this environment while keeping the rest of the stack version-flexible.
tensorflow>=2.16.0,<3.0.0
transformers>=4.30.0,<5.0.0
scikit-learn>=1.3.0,<2.0.0
scikit-fuzzy==0.4.2
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
pyyaml>=6.0.0,<7.0.0
tqdm>=4.65.0,<5.0.0
pytest>=7.4.0,<9.0.0
pytest-asyncio>=0.21.0,<1.0.0
black>=23.0.0,<25.0.0
flake8>=6.0.0,<8.0.0
scipy>=1.10.0,<2.0.0
joblib>=1.3.0,<2.0.0
imbalanced-learn>=0.11.0,<1.0.0
requests>=2.28.0,<3.0.0
